refactor:调整icon组件代码和结构
